refactor: Restructure messaging package, rename classes for clarity, and add MetadataResponseType
