+++ mossy-password/src/main/kotlin/pl/dayfit/mossypassword/messaging/resolver/DeleteMessageHandler.kt
@@ -1,31 +1,34 @@
 package pl.dayfit.mossypassword.messaging.resolver
 
+import messaging.VaultRequestMessageDto
+import messaging.VaultResponseMessageDto
+import messaging.request.type.DeletePasswordRequestType
+import messaging.response.type.DeletePasswordResponseType
 import org.springframework.stereotype.Component
-import pl.dayfit.mossypassword.dto.vault.VaultRequestMessageDto
-import pl.dayfit.mossypassword.dto.vault.VaultResponseMessageDto
-import pl.dayfit.mossypassword.dto.vault.request.DeletePasswordRequest
-import pl.dayfit.mossypassword.dto.vault.response.DeletePasswordResponse
 import pl.dayfit.mossypassword.service.VaultMessagingService
 import java.util.concurrent.CompletableFuture
+import kotlin.reflect.KClass
 
 @Component
 class DeleteMessageHandler(
     private val vaultMessagingService: VaultMessagingService
-) : AbstractMessageHandler<DeletePasswordRequest, DeletePasswordResponse>() {
+) : AbstractMessageHandler<DeletePasswordRequestType, DeletePasswordResponseType>() {
     companion object {
         private const val TOPIC = "delete"
     }
 
-    override fun handleMessage(message: VaultRequestMessageDto<DeletePasswordRequest>): CompletableFuture<VaultResponseMessageDto<DeletePasswordResponse>> {
+    override fun handleMessage(message: VaultRequestMessageDto<DeletePasswordRequestType>): CompletableFuture<VaultResponseMessageDto<DeletePasswordResponseType>> {
         vaultMessagingService.sendMessageToTopic(
             TOPIC,
             message
         )
 
-        val future = CompletableFuture<VaultResponseMessageDto<DeletePasswordResponse>>()
-        pending[message.correlationId] = future
+        val future = CompletableFuture<VaultResponseMessageDto<DeletePasswordResponseType>>()
+        pending[message.correlationId.toString()] = future
         return future
     }
 
-    override fun supportedType() = DeletePasswordRequest::class
+    override fun doSupport(type: KClass<*>): Boolean {
+        return type == DeletePasswordRequestType::class
+    }
 }
