+++ mossy-password/src/main/kotlin/pl/dayfit/mossypassword/service/VaultCommunicationService.kt
@@ -2,8 +2,8 @@ package pl.dayfit.mossypassword.service
 
 import messaging.VaultRequestMessageDto
 import messaging.VaultResponseMessageDto
-import messaging.type.AbstractVaultRequestType
-import messaging.type.AbstractVaultResponseType
+import messaging.request.type.AbstractVaultRequestType
+import messaging.response.type.AbstractVaultResponseType
 import org.springframework.amqp.rabbit.annotation.RabbitListener
 import org.springframework.amqp.rabbit.core.RabbitTemplate
 import org.springframework.amqp.support.AmqpHeaders
