+++ mossy-vault-shared/src/main/kotlin/messaging/request/type/AbstractVaultRequestType.kt
@@ -1,4 +1,4 @@
-package messaging.type
+package messaging.request.type
 
 abstract class AbstractVaultRequestType {
 }
