+++ mossy-vault-shared/src/main/kotlin/messaging/request/type/CiphertextResponseType.kt
@@ -1,0 +1,6 @@
+package messaging.request.type
+
+import messaging.response.type.AbstractVaultResponseType
+
+class CiphertextResponseType : AbstractVaultResponseType() {
+}
