+++ mossy-vault-shared/src/main/kotlin/messaging/request/type/PasswordSaveRequestType.kt
@@ -1,4 +1,4 @@
-package messaging.type
+package messaging.request.type
 
 import type.PasswordSaveType
