+++ mossy-vault-shared/src/main/kotlin/messaging/response/type/AbstractVaultResponseType.kt
@@ -1,4 +1,4 @@
-package messaging.type
+package messaging.response.type
 
 abstract class AbstractVaultResponseType {
 }
