+++ mossy-vault-shared/src/main/kotlin/messaging/response/type/DeletePasswordResponseType.kt
@@ -1,8 +1,8 @@
-package messaging.type
+package messaging.response.type
 
 import java.util.UUID
 
-data class DeleteVaultResponseType(
+data class DeletePasswordResponseType(
     private val domain: String,
     val passwordId: UUID
 ) : AbstractVaultResponseType()
