+++ mossy-vault-shared/src/main/kotlin/messaging/response/type/PasswordQueryResponseType.kt
@@ -1,4 +1,4 @@
-package messaging.type
+package messaging.response.type
 
 class PasswordQueryResponseType : AbstractVaultResponseType() {
 }
