--- mossy-vault-shared/src/main/kotlin/messaging/type/CiphertextResponseType.kt
@@ -1,4 +1,0 @@
-package messaging.type
-
-class CiphertextResponseType : AbstractVaultResponseType() {
-}
